Don't attempt to copy files which already exist

+++ utility.py
@@ -56,7 +56,8 @@ def copy_to_work_area(input_file_list, work_area):
     outputs = []
     for ms_name in input_file_list:
         output_name = os.path.join(work_area, os.path.basename(ms_name))
-        copytree(ms_name, output_name)
+        if not os.path.exists(output_name):
+            copytree(ms_name, output_name)
         outputs.append(output_name)
     return outputs
